fix bugs

+++ presentation/src/main/java/com/cheesecake/presentation/screens/favorite/favoriteLeagues/FavoriteLeagueUIState.kt
@@ -7,6 +7,6 @@ data class FavoriteLeagueUIState(
     val leagueName: String = "",
     val leagueCountry: String = "",
     val onFavorite: () -> Unit,
-    val onFavoriteLeagueClick: (Int) -> Unit,
+    val onFavoriteLeagueClick: () -> Unit,
     val isFavourite: Boolean = false
 )
